Seed bug
A few new PlantTypes that should produce seeds and are marked to do so still do not.

This is why

- public bool PollenProducing => Plant.IsCrossable && Plant.PlantStatus >= PlantStatus.FullGrownPlant;

old method

if (Pollinated && LeftSeeds > 0 && Plant.Reproduces)

So since these new types only create seeds but can't be pollinated per https://uo.com/wiki/ultima-online-wiki/gameplay/crafting/plant-growing/fragrant-and-peculiar-seeds/, the system stops them from producing seeds because for it to pass the pollen check it would have to have crossable set to true, which it is correctly false. The change just bypasses the check for those types, not the most elegant but it works.