Remove Linq
- https://github.com/TrueUO/TrueUO/blob/e62b39f1c21e9310fdcfd1d6208ed25ba562115b/Scripts/Misc/Notoriety.cs#L495

I get that Any is better than other uses of LINQ but there is still the overhead.

Also here we are running a linq inside a linq, before this change, once pets are involved.
https://github.com/TrueUO/TrueUO/blob/e62b39f1c21e9310fdcfd1d6208ed25ba562115b/Scripts/Misc/Notoriety.cs#L495

Here we are returning to the CheckAggressor, which currently is LINQ, prior to this.
return source.AllFollowers.Any(follower => CheckAggressor(follower.Aggressors, target));