VendorSellBug
When selling items to an NPC they simply go into the BuyPack to allow them to be "re purchased" by other players. The problem is this BuyPack and items sold NEVER decay. They stay in that bag. If that bag becomes full of hundreds of items when players go to vendor buy it will freeze / cause a delay on their end while all those items are loaded.

On EA I noticed that when I sold a random item, about 10 minutes later that item was no longer there. I am guessing they do a sort of clean up.

Now there are still a few issues:

One if the item decays while waiting to be bought, the client will crash. Second is the displaying of time left.

Here I simply add a Decaying Item Socket to non-resource/stacking items when sold and another check to remove them when purchased. This is not intended for a production shard but more to showcase one way we could potentially fix this issue.