VendorSellCheckBuyPack
Whenever you sell items to an NPC it goes to the BuyPack. After 10 minutes it becames "deletable" and will be deleted the next time someone uses the Buy option.

The problem is if someone, especially on an "off the beaten path" npc, only SELLS items this call to delete items in the BuyPack never happens. It can add hundreds/thousands of items over time.

This simply adds another check everytime someone goes to sell items.